Support symlinks in `levels` through workaround (`levels-parent/levels`)
Note: in the following text, `levels` refers to the `levels`
directory inside the /assets/ file (`data.zip`), or the `levels`
subdirectory inside the "basedir" (`-basedir` program option) directory,
depending on the context in which this term is used.

This changeset introduces support for a setup that can address the
following use cases:

  1. The `levels` "basedir" subdirectory is a symlink to a directory
    that the user has under source control, for example to allow sharing
    those levels across several computers, and/or with other community
    members, etc.;

  2. `vvvvvvvv` is run with `-basedir` specifying a user-specific
    "profile" directory containing a user-specific `saves` subdirectory,
    but a "shared" `levels` (symlink to a) directory is desirable;

    1. In a household sharing a computer (gaming) account amongst
      multiple users wanting to share `vvvvvvvv` user levels seamlessly,
      this allows for a "natural" setup;

The setup catering for the use cases stated above would "naturally" use
a `{basedir}/levels` symlink pointing to the directory containing the
user levels.

Previously, this setup could not be made to work (in a "natural"-ish?)
way.

`physfs` does not support symlinks to directories to be treated as
regular directories when dealing with "enumerating" their entries
(`PHYSFS_enumerate()`).

Previously, this meant that `vvvvvvvv` could not have a working symlink
to a `levels` subdirectory, as there would be no `levels` entries being
processed by `PHYSFS_enumerate()`.

This change implements a workaround in "user" (game) code by allowing
`levels` to be symlinked through a symlink to its parent directory
(`levels-parent`), and treating entries from those directories in the
same way as entries in regular `levels` (sub/)directories, allowing
level resources to be correctly found and processed.

Typically, users wanting to use the "symlinked levels" subdirectory
would normally do something like this:

    $ cd {basedir}
    $ ln -vins {real_levels_directory's_parent} levels-parent
    $ ln -vins levels-parent/levels .

So, for example, in a "standard" (no "-basedir") user directory on *nix
(Linux and others), the user might do something like this:

    $ cd ~/.local/share/VVVVVV
    $ ln -vins {real_levels_directory's_parent} levels-parent
    $ ln -vins levels-parent/levels .

This workaround is (should be) 100% backwards-compatible, and should not
break any typical setups users might already have.

The level editor still uses the `levels` directory as before, but this
also worked previously in the case where `levels` is a symlink to a
directory, so no code changes specific to the level editor have been
made in this changeset.

What this changeset mainly addresses is the functionality around playing
user levels, loading, saving and deleting "level saves", and finding and
loading resources associated to user levels.

As the environment for this "symlink to `levels`" setup is likely to be
*nix only, this workaround is disabled for Windows builds.

In order to support the extra case of levels being accessible through
`*/levels` (potentially symlinked) directories, the code has been made
more flexible in a few places, abstracting the "get basename" concept to
a new function `FILESYSTEM_getBaseName()`, and using standard functions
to extract pathname components, instead of assuming fixed string
offsets.

--

This commit has the first working implementation, tested with and
without each of these separate cases:

  * no `levels` directory present (first run, for example);

    * NOTE: directory gets created by `vvvvvvvv`, as before;

  * `levels` being a regular directory;

    * NOTE: directory gets correctly enumerated, and its levels
      processed, as before;

  * `levels` being a symlink to a directory;

    * NOTE: levels from this symlinked directory are still not
      processed (as was the case before this change), as that is a
      `physfs` limitation.

  * `levels-parent` being a symlink to a directory containing a `levels`
    subdirectory;

    * NOTE: levels from a regular directory `levels-parent/levels` are
      processed as if they were located in a regular `levels` directory;

  * `levels-parent` being a regular directory containing a `levels`
    subdirectory;

    * NOTE: same as the case just above this one;

--

Status:

  * The code in this commit is fully functional, but there are still
    "left-overs", and possibly a few minor changes to be done before a
    final "pull request" can be made.